Rename class name from Authority to Role
