+++ src/main/java/com/kat/bachaat/model/Role.java
@@ -1,0 +1,41 @@
+package com.kat.bachaat.model;
+
+import javax.persistence.*;
+
+@Entity
+@Table(name="tbl_role")
+public class Role {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "role_id", updatable = false, nullable = false)
+    private int roleId;
+    @Column(name="role_name")
+    private String name;
+
+    public Role() {
+    }
+
+    public int getRoleId() {
+        return roleId;
+    }
+
+    public void setRoleId(int roleId) {
+        this.roleId = roleId;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @Override
+    public String toString() {
+        return "Role{" +
+                "roleId=" + roleId +
+                ", name='" + name + '\'' +
+                '}';
+    }
+}
